==> notebooks/02_cleaning.ipynb <==
# LIMPIEZA DE DATOS ECOMMERCE

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import plotly.graph_objects as go

import sqlalchemy as sa

import re

pd.options.display.float_format = '{:.2f}'.format


%config IPCompleter.greedy=True # Activa el autocompletar avanzado en Jupyter

from IPython.display import display, HTML, Image
display(HTML("""<style> th{ text-align: center !important; font-weight: bold !important; } td{ text-align: center !important; }</style>"""))

---
## ENTENDIENDO EL SIGNIFICADO DE LAS VARIABLES
---

| Variable      	| Tipo de Dato 	| Descripción                                                                             	|
|---------------	|--------------	|-----------------------------------------------------------------------------------------	|
| index         	| BIGINT       	| Índice único de cada fila.                                                              	|
| event_time    	| TEXT         	| Fecha y hora del evento en formato YYYY-MM-DD HH:MM:SS UTC.                             	|
| event_type    	| TEXT         	| Tipo de evento: view (vista de producto), cart (añadido al carrito), purchase (compra). 	|
| product_id    	| BIGINT       	| Identificador único del producto.                                                       	|
| category_id   	| BIGINT       	| Identificador único de la categoría del producto.                                       	|
| category_code 	| TEXT         	| Nombre de la categoría del producto (puede ser NULL).                                   	|
| brand         	| TEXT         	| Nombre de la categoría del producto (puede ser NULL).                                   	|
| price         	| FLOAT        	| Precio del producto en la moneda base.                                                  	|
| user_id       	| BIGINT       	| Identificador único del usuario.                                                        	|
| user_session  	| TEXT         	| Identificador único de la sesión del usuario.                                           	|

---
## IMPORTACION DE DATOS
---

df = pd.read_csv("../data/originals/ecommerce.csv")
df

---
## ANÁLISIS DEL DATAFRAME
---

# Información general del DataFrame
print(f"- Información general del DataFrame:")
df.info()

# Visualizamos la cantidad de filas y columnas
print(f"\n- El DataFrame consta de: {df.shape[0]} filas y {df.shape[1]} columnas")

# Visualizamos el índice del DataFrame
print(f"- El índice del DataFrame es: {df.index}")

# Visualizamos el nombre de las columnas
print(f"- El nombre de las columnas es: {df.columns.tolist()}\n")

# Visualizamos los registros duplicados
print(f"- Registros duplicados: {df.duplicated().sum()}\n")

# Visualizamos los valores unicos de cada columna
print(f"- Valores únicos por columna:\n{df.nunique()}\n")

# Visualizamos los valores nulos por columna
print(f"- Valores nulos por columna:\n{df.isnull().sum()}\n")

---
## FORMATEO Y CORRECCIÓN DE NOMBRES DE VARIABLES
---
- La columna index es un indice que viene con los datos que no vamos a utilizar por lo que se puede eliminar
- La columna category_code la podemos eliminar ya que la mayoria de registros son nulos y ya tenemos category_id
- La columna brand (nombre de la categoria del producto) la podemos eliminar ya que tiene muchos nulos, hace referencia al nombre de la categoría del producto y ya utilizamos el category_id
- Formatear y renombrar variables

# Eliminar las columnas 'index', 'category_code' y 'brand'
df = df.drop(columns=['index', 'category_code', 'brand'])

# Limpiar y estandarizar nombres de columnas
df.columns = [re.sub(r'[^\w\s]', '', col).lower().strip(' _').replace(' ', '_') for col in df.columns]

# Renombramos las variables
df.rename(columns={
    'event_time': 'fecha',
    'event_type': 'evento',
    'product_id': 'producto',
    'category_id': 'categoria',
    'price': 'precio',
    'user_id': 'usuario',
    'user_session': 'sesion'
}, inplace=True)
df

---
## CONVERTIR TIPOS DE DATOS
---
- Convertir 'fecha' en tipo de dato datetime64
- Convertir la variable 'evento' a tipo category ya que tienen pocos valores únicos, esto optimizará memoria y mejorará la velocidad en análisis y modelos de Machine Learning

# Convertir 'fecha' a datetime64 y eliminar el componente de zona horaria (UTC) si existe
df['fecha'] = pd.to_datetime(df['fecha'], errors='coerce').dt.tz_localize(None)

# Convertir la variable 'evento' a tipo category
df['evento'] = df['evento'].astype('category')

# Verificar los cambios
df.info()
display(df)

---
## ANÁLISIS DE NULOS
---
- 'sesion' identifica la sesion de navegacion de los usuarios, eliminamos los nulos ya que es una variable relevante y no tiene sentido hacer ningún tipo de imputación

# Visualizamos los valores nulos por columna
print(f"- Valores nulos por columna:\n{df.isnull().sum()}\n")

# Número de registros antes de eliminar nulos
print(f"Registros antes de eliminar nulos: {df.shape[0]}")

# Eliminar los registros con valores nulos (en cualquier columna, pero asumimos que solo 'sesion_usuario' tiene nulos)
df = df.dropna()

# Número de registros después de eliminar nulos
print(f"Registros después de eliminar nulos: {df.shape[0]}")

# Visualizamos los valores nulos por columna
print(f"\n- Valores nulos por columna:\n{df.isnull().sum()}")

---
## ANÁLISIS DE VARIABLES CUANTITATIVAS
---
Conclusiones:

* Existen valores negativos en la variable 'precio', puede ser por devoluciones pero no deberia existir un precio negativo en el producto
* Los registros de los productos donde el precio del valor es menor o igual a cero no se concentran en un solo producto


Acciones:

* eliminar los registros donde el precio del producto es igual o menor a cero

# Número de registros con precio <= 0
print(f"- Cantidad de registros con precio menor o igual a 0: {df[df.precio <= 0].shape[0]}") 

# Frecuencia de los productos con precio <= 0
print(f"\nFrecuencia de los productos con precio <= 0: {df[df.precio <= 0].producto.value_counts().head(10)}")

print(f"\n- Número de registros antes de eliminar precios <= 0: {df.shape[0]}")

# Eliminar los registros donde el precio es igual o menor a cero
df = df[df.precio > 0]

print(f"- Número de registros después de eliminar precios <= 0: {df.shape[0]}")

---
## GUARDAR EL DATAFRAME EN FORMATO PICKLE
---

# Guardar el DataFrame en formato pickle
df.to_pickle("../data/processed/ecommerce_cleaning.pkl")
print("El DataFrame ha sido guardado en 'data/processed/' con el nombre 'ecommerce_cleaning.pkl'")

# Guardar el DataFrame en formato CSV
df.to_pickle("../data/originals/ecommerce_cleaning.csv")
print("El DataFrame ha sido guardado en 'data/originals/' con el nombre 'ecommerce_cleaning.csv'")
