==> notebooks/03_transformation.ipynb <==
# TRANSFORMACIÓN DE DATOS

---
## IMPORTAR PAQUETES 
---

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

import sqlalchemy
import janitor

import re

pd.options.display.float_format = '{:.2f}'.format

%config IPCompleter.greedy=True # Activa el autocompletar avanzado en Jupyter

import warnings
# Ignora cualquier FutureIncompatibilityWarning antes de importar holidays
warnings.filterwarnings("ignore", category=UserWarning)
warnings.filterwarnings("ignore", category=FutureWarning)

# Ignora todo tipo de warning antes de cargar holidays
warnings.simplefilter("ignore")

import holidays

from IPython.display import display, HTML
display(HTML("""<style> th{ text-align: center !important; font-weight: bold !important; } td{ text-align: center !important; }</style>"""))

---
## IMPORTAR DATOS
--- 

# Cargar datos
df = pd.read_pickle("../data/processed/ecommerce_cleaning.pkl")
df

---
## CREACIÓN DE NUEVAS VARIABLES
---
- En este tipo de contexto o sector como ventas existe mucha influencia de la variable tiempo por lo que son importantes las fechas:
- Extraer componentes de la fecha
- Crear variables de calendario (festivos locales)
- Indicadores exógenos, se utiliza para analizar series temporales, para recoger eventos pasados que se pueden medir en una nueva variable, son variables externas que no forman parte de la serie en estudio pero que pueden influir en su comportamiento. Se usan para mejorar la precisión de los modelos de predicción al incorporar factores externos relevantes. (Black Friday, San Valentin...)

### Variables de calendario: Festivos locales (Rusia)
- Para incorporar festivos podemos usar el paquete holidays.
- Lo instalamos con: conda install -c conda-forge holidays
- Lo importamos con: import holidays

Y podemos ver el listado de países y el uso básico en:
https://github.com/dr-prodigy/python-holidays

# Extraer componentes de la fecha en nuevas columnas
df['date'] = df['fecha'].dt.date
df['año'] = df['fecha'].dt.year
df['mes'] = df['fecha'].dt.month
df['dia'] = df['fecha'].dt.day
df['hora'] = df['fecha'].dt.hour
df['minuto'] = df['fecha'].dt.minute
df['segundo'] = df['fecha'].dt.second

# Establecer la columna 'fecha' como índice
df = df.set_index('fecha')

# Cambiar la variable date a tipo datetime64
df['date'] = pd.to_datetime(df['date'])

# Accedemos a los festivos locales de Rusia
festivo_ru = holidays.RU(years = 2020)
print("Días festivos en Rusia:")
for fecha, nombre in festivo_ru.items():   
    print(f"{fecha}: {nombre}")

# Creamos una variable que diga en cada registro si es un día festivo o no (si = 1, no = 0)
df['festivo'] = df['date'].apply(lambda x: 1 if x in festivo_ru else 0)

# Registros que existen en cada fecha festiva 
print(f"\nRegistros en cada fecha festiva: \n{df.query('festivo == 1').date.value_counts().sort_index()}")

# Filtrar las filas donde la columna 'festivo' sea igual a 1
print(f"\nRegistros en día de fiesta:")
display(df.query('festivo == 1'))


### Indicadores exógenos: Días no necesariamente festivos pero con interés comercial: Black Friday, Cyber Monday, Reyes, San Valentin
- Vamos a añadir indicadores para Black Friday y San Valentín

# 2019-11-29 = black friday en Rusia, creamos nueva columna y los registros que coincidan con la fecha los cambiamos a 1 en la variable
df['black_friday'] = 0
df.loc['2019-11-29','black_friday'] = 1

# 2020-02-14 = San Valentin en Rusia, creamos nueva columna y los registros que coincidan con la fecha los cambiamos a 1 en la variable
df['san_valentin'] = 0
df.loc['2020-02-14','san_valentin'] = 1

print(f"{df['black_friday'].value_counts()}\n")
print(f"{df['san_valentin'].value_counts()}")

---
## Reordenar variables
---

# Extraemos las variables en una lista
variables = df.columns.to_list()

# Reordenamos las variables para que queden en el orden deseado
orden = ['usuario', 'sesion', 'categoria', 'evento', 'producto', 'precio']

# Extraemos el resto de las variables
resto_var = [nombre for nombre in variables if nombre not in orden]

# Reordenamos el DataFrame
df = df[orden + resto_var]
df

df.info()

---
## GUARDAR EL DATAFRAME EN FORMATO PICKLE
---

# Guardar el DataFrame en formato pickle
df.to_pickle("../data/processed/ecommerce_transformado.pkl")
print("El DataFrame ha sido guardado en 'data/processed/' con el nombre 'ecommerce_transformado.pkl'")

# Guardar el DataFrame en formato CSV
df.to_csv("../data/originals/ecommerce_transformado.csv")
print("El DataFrame ha sido guardado en 'data/originals/' con el nombre 'ecommerce_transformado.csv'")

conexion a .sql

from sqlalchemy import create_engine
df = pd.read_pickle('../data/processed/ecommerce_transformado.pkl')
engine = create_engine('mysql+pymysql://root:MySQL.2025@localhost/ecommerce')
df.to_sql('ecommerce_transformado', con=engine, index=False, if_exists='replace')
